backend/requirements.txt
fastapi>=0.100.0
# uvicorn[standard]
# motor                   # Async driver for MongoDB
# beanie                  # ODM for MongoDB
pydantic-settings       # For managing .env files
# python-multipart        # For file uploads (OCR images)
# python-jose[cryptography] # For JWT Auth
# passlib[bcrypt]         # Password hashing
# langchain               # AI Logic
#   # Google Gemini wrapper
opencv-python-headless  # Computer Vision (Server version)
#            # OCR (Warning: Large install)
#               # OCR wrapper
python-multipart        # For handling multipart/form-data (file uploads)
#           # For loading environment variables from .env files
# langchain_community
python-dotenv>=1.0.0
pymongo>=4.6.0
motor>=3.3.0
beanie>=1.20.0
pydantic>=2.5.0
langchain-google-genai>=0.0.10
pytesseract>=0.3.10
Pillow>=10.0.0
unidecode>=1.3.6